internal/cmd/start_test.go
// start_test.go
package cmd_test

import (
	"testing"
	"time"

	"github.com/guarzo/canifly/internal/cmd"
	"github.com/stretchr/testify/assert"
)

// TestStart checks if Start runs without errors when environment variables are set properly.
// Note: This won't fully test server startup because that would block indefinitely.
// If necessary, consider refactoring Start to allow dependency injection or mock out parts of the code.
func TestStart(t *testing.T) {
	// Set required environment variables that LoadConfig and services might expect
	t.Setenv("EVE_CLIENT_ID", "test_id")
	t.Setenv("EVE_CLIENT_SECRET", "test_secret")
	t.Setenv("EVE_CALLBACK_URL", "http://localhost/callback")

	// If your LoadConfig relies on other env vars or config files, set them here too.
	// For example:
	t.Setenv("SECRET_KEY", "c2VjcmV0a2V5MTIz") // base64 of 'secretkey123'
	t.Setenv("PORT", "8888")

	// Now call Start. Because Start will attempt to run a server, this will block.
	// In a real unit test, we might want to mock runServer or at least ensure it doesn't block indefinitely.
	// One approach: run Start in a goroutine and just ensure it doesn't panic immediately.
	// Another approach: temporarily modify Start to return after setup for testing.

	done := make(chan error, 1)
	go func() {
		err := cmd.Start()
		done <- err
	}()

	// Wait a short time and then assume success if no error returned
	// In reality, Start() will block on ListenAndServe, so this test might time out.
	// A better approach: factor runServer out further or mock the HTTP server.

	// We'll just test that it at least doesn't fail immediately on loading config/services.
	// If you need to actually verify the server started, you'd do more here, like try connecting to it.

	select {
	case err := <-done:
		// If we get here, that means Start returned immediately, which it shouldn't unless there's an error.
		assert.NoError(t, err)
	case <-time.After(500 * time.Millisecond):
		// If half a second passed and no error, we assume it started okay.
		// Realistically, you might send a signal here or mock runServer.
	}
}
